server/requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
dj-email-url==1.0.1
dj-rest-auth==2.1.2
Django==3.1
django-allauth==0.43.0
django-cache-url==3.1.2
django-cors-headers==3.5.0
django-ipware==3.0.1
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
environs==8.0.0
idna==2.10
iniconfig==1.0.1
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
marshmallow==3.7.1
more-itertools==8.5.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
phonenumbers==8.12.11
Pillow==8.0.1
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-asyncio==0.14.0
pytest-django==3.9.0
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
simplejson==3.17.5
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.11
